added ship specialized hold checks and capacitys mining laser will dump ore into ore hold, if ship has one salvager will dump salvage into salvage hold, if ship has one updated integer problem in GetPackagedVolume() method add check/method to create mined ore items (cat::Asteroid) (error fix) added check for moving stacks of items using "loot all" button (error fix) added capacity checks for specalized PI holds

+++ src/eve-server/inventory/InventoryItem.h
@@ -153,7 +153,7 @@ public:
     static uint32           CreateTempItemID( ItemData &data);
     /* loads attributes for this item */
     //bool LoadAttributes();
-    uint32                  GetPackagedVolume();
+    double GetPackagedVolume();
 
     /* specific funtions for ShipItem, virtual here to allow generic class access */
     virtual void            SetPlayer(Client* pClient)  { /* do nothing here */ }
